refactor: simplify import path for fetchRecents service
