+++ src/views/Recents/store/fetchRecentsThunk.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi, beforeEach, Mock } from 'vitest';
 import { fetchRecentsThunk, fetchRecentsThunkExtraReducers } from './fetchRecentsThunk';
-import { fetchRecents } from '../services/fetchRecents';
+import { fetchRecents } from '../services';
 import configService from 'app/core/services/config.service';
 import { excludeHiddenItems, getItemPlainName } from 'app/crypto/services/utils';
 import { storageActions } from 'app/store/slices/storage';
